config: assert that config is loaded before using
When constructing the Config class, the configuration dict is not
immediately available. Instead, it needs to be build-up by iteratively
calling find_missing_repos.

This behavior has shown error-prone for developers that assume they can
already use any Config.get_* method without running the repo import
loop. As it also cannot easily be changed without a major refactoring,
we just add an internal assertion that checks if the config is loaded
before any operation on it.

As we now have the assertion, we also need to replace the useless call
to _get_repo_dict() in the constructor by the static empty-dict value.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>