dump: drop overrides when resolving refs
When running the dump plugin with --resolve-refs, all references are
resolved and added to the commit of the corresponding repo. By that,
this information is identical with a potential entry in the overrides
section. Further, we do not want to have override entries for not
defined repositories, as the configuration should be fully resolved.

We now drop the overrides section completely, when running the dump
plugin with --resolve-refs.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>