gnupg: do not spam console on debug logging
When running kas with debug logging, we configure the root logger. By
that, all loggers that are later created will also inherit that log
level. Usually this is fine, but gnupg is quite verbose when logging in
debug mode. This visually interferes with the kas log messages.

As the gnupg part is anyways only used by kas internally, we just limit
the logging to INFO for the gnupg logger.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>