Dockerfile: install python3-cairo
The python3-cairo is required to run the oe-core pybootchartgui scripts.
The oe-core buildstats collect a lot of information during the build but
this output is hard to read and understand. This pybootchartgui script
process the logs produced by the buildstats and generates some pretty charts.

The installation will increase the image size 10Mb

| $ docker inspect ghcr.io/siemens/kas/kas:latest --format='{{.Size}}'
| 1132649238
| $ docker inspect kas-local --format='{{.Size}}'
| 1142455807

We can test it with:

  kas-container shell ...
  echo 'INHERIT += "buildstats"' >> conf/local.conf
  bitbake core-image-minimal
  $(find /work/ | grep scripts/pybootchartgui/pybootchartgui.py) --output=buildchart tmp/buildstats

The '--output=' argument is mandatory otherwise the script will try to open a gui
and fail with: ModuleNotFoundError: No module named 'gi'.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>