Add recursive check to change ownership if ssh directory is present for docker rootless containers
 Currently attempting to use the "--ssh-dir" option will import a
directory into a docker container that is still owned by the root user
of the container.

This does not work as "builder" cannot access these files making this
option useless in a rootless environment. This PR adds a small change to
check for the existence of the ssh directory if it exist and changes its
ownership to the appropriate user.

Signed-off-by: Aidan Moss <aidan.moss@altec.com>
[Jan: whitespace fix, implant commit message from original PR]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>