includehandler: improve reporting of yaml format errors
When passing a yaml file that is incorrectly formatted, the parser
raises an error. This error needs to be caught, as otherwise a huge
splat is given back to the user which hides the error.

We now handle this case and report the line at which the error most
likely occurred.

Reported-by: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>