tests: do not make assumptions about KAS_BUILD_DIR or KAS_WORK_DIR
The tests currently assume, that these environment variables are unset.
This holds, as we explicitly clear them before test execution. However,
it makes it impossible to test scenarios where these variables are set.

As a preparation to parameterize some of the tests, we modify all tests
to not make any assumptions about these directories. We further document
this in the test writing guide.

No functional change.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>