kas-container: pass script version to kas in container
As a preparation to establish a stable interface between kas-container
and kas, we pass the kas-container script version as an environment
variable into the container. There it can be evaluated by kas.
We further terminate the container with a meaningful error message in
case a kas-container call is detected but kas inside the container does
not have a matching command.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>