schema: enforce allowed values of layer property
The layer property has a set of well-known values to "disable" the layer
(i.e. not add it to the bblayers). This is currently checked in kas, but
we better enforce that via the schema as then IDEs can help the user to
set correct values. Further, incorrect values are now rejected, while
they were ignored before. As the documentation currently is imprecise
about the types of the allowed values (e.g. 0 vs "0"), we also allow the
int and boolean types provided as string.

No functional change (w.r.t valid configurations).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>