deprecate various magic values to disable a layer
Instead of allowing a variety of strings to disable a layer, we stick to
a single value "disabled". By that, we only provide a single way to
disable a layer, which both helps the IDEs, as well as avoids ambiguity.
As we use jsonschema draft-04, we cannot use the "const" property, so
instead we just use an enum with a single value.

To give downstream users some time to migrate, we still allow the old
values, but file a deprecation warning if used.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>