lazily initialize key handlers for repo verification
Previously, all registered keyhandlers were always initialized, even if
no repo was signed with a key managed by the handler. By that, the
keyhandler dependencies were mandatory (despite marked as optional).
Also, the key handler directories were always created and in case of
GnuPG a decent amount of python code was executed.

We now change that by only initializing the keyhandlers that are
registered using the signers config property. As the number of
single-letter variables for the handler config is now bigger than
manageable, we also give them telling names.

Fixes: 5176c9d52 ("add KeyHandler class to work with local keystores")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>