Dockerfile: bind mount sources dir instead of copy
Previously we copied the repo as context into the container, installed
kas and later deleted the context again. By that, an additional layer is
created that potentially leaks data from the context into the image and
also wastes (a bit) of space. We fix this by bind-mounting the dir
instead. As pip generates temporary data in that dir, we need to mount
it rw (the written data is discarded).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>