docs: add note on YAML support and its limitations
As kas internally uses PyYAML to parse YAML documents, only version 1.1
of the spec is supported. Further, the parser does not correctly handle
non string keys of mappings (they are parsed as-is instead of being
converted to strings on parsing), violating Section 8.2.1 of the YAML
1.1 spec.

We document these limitations in the project configuration along with
workarounds.

Reported-by: Cedric Hombourger <cedric.hombourger@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>