kas-container: clear KAS_FILES before parsing command arguments
The command line parser takes each file argument and appends it
to its global KAS_FILES variable. If that variable happens to
be set/exported in the calling shell, kas will be called with
a space-separated list of kas files (when : is expected when
multiple files should be parsed). Since KAS_FILES is not an
input/documented environment variable [1], clear it before we
parse the kas-container command line.

[1] https://kas.readthedocs.io/en/latest/command-line.html#environment-variables

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>