docs: precisely describe where SHELL env var is in effect
Previously the documentation states that both the SHELL and TERM
variables are only forwarded into the kas shell plugin. This was not
true as they are always forwarded into the bitbake environment (but not
into the bitbake setup script). While this does not matter for normal
bitbake task execution (which is always executed under /bin/sh), it
does matter for the execution of the devshell.

We now align the documentation with the implementation and precisely
state into which environments the variables are forwarded.

No functional change.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>