kas-container: fallback to xterm if TERM is not supported
Currently, we just forward the TERM variable into the container. This
breaks, if the user uses a terminal emulator on the host that misses the
terminfo inside the container (like foot). Then, tools like kas menu or
bitbake fail to start.

To mitigate this, we check in the container entrypoint if we have the
terminfo for that terminal. If not, we fall back to xterm, as nearly all
terminal emulators implement a superset of xterm.

Reported-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>