kas-container: propagate host timezone via entrypoint
Bind-mounting /etc/timezone into the container overwrites the symlinks
target, as both docker and podman mount without the AT_SYMLINK_NOFOLLOW
flag.  This breaks yocto builds which rely on correct information in
/usr/share/zoneinfo/*. To fix this, we pass the host timezone via an
environment variable to the container and patch the timezone in the
entrypoint. By that, we also preserve the difference between the TZ
variable and the /etc/timezone, /etc/localtime data.

Fixes: 378366395 ("kas-container: propagate timezone information ...")
Closes: #162
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>