GitLab CI: set safe.directory exception for git outside of kas
In 70c98b32f we moved the git safedir handling from the container
entrypoint to the kas gitconfig setup, where already other CI related
aspects are handled. This is fine for git calls from inside the kas
environment, however it breaks git calls outside of kas, as there the
kas gitconfig is not in effect.

We fix it by a partial revert of the workaround in d9b6a7798. Instead of
completely disabling the safedirs security feature, we now only disable
it for the CI_PROJECT_DIR.

Fixes: d9b6a7798 ("Revert container: Disable git safe.directory ...")
Reported-by: Gokhan Cetin <gokhan.cetin@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>