clean: avoid creation of tmp data to recover from full-disk scenario
When running kas clean with a provided config file, kas needs to setup
the environment to process that config. This requires some temporary
space, which is not available on a full disk. By that, the clean fails
and makes it impossible to recover from that situation (at least with
kas mechanisms).

We fix this by only parsing the top-level config file to get the build
system on clean. This is done without the need for temporary storage.
On purge, we still parse the full config, as we need the repository
information. As the clean task is run beforehand (due to the
inheritance), we should have enough space to setup the tmpdirs and
perform the repo cloning.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
[Jan: add code comment about first config file]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>