Allow branch override
Since eb4d112 when commit and branch are specified, kas checks if the
commit is actually contained in the branch.
This makes it impossible to have a branch set for a repo but use a
commit override to build a merge request on this repo (obviously not
yet on the specified branch).

This change allows the branch property of repos to be overridden to a
specific value, or even to Null.

Reviewed-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Nicolas Labriet <nlabriet@centralp.fr>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>