fix: do not raise on task exceptions in shutdown cleanup
When shutting down the event loop, we are already in a termination
context. In this context, we must not raise any exceptions to not
interact with the event loop again. Otherwise the eventloop is not shut
down properly leading to exceptions thrown in non-raisable contexts.

Handling these exceptions is anyways not needed, as the root cause for
the cancellation (an exception) is already handled in the main execution
of the event loop.

Fixes: f4beea37f ("graceful cancellation of external commands")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>