repos: only fetch on update if commit is not pinned
In case the checkout is executed with --update and the commit of a repo
is pinned (directly via the config, not via a lockfile), we can safely
skip the fetching if we already have that commit. This slightly speeds
the checkout up as no irrelevant fetch is executed in this case.

Closes: #164
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>