test: add test for failed fetching of repos
We currently only test the good case of fetching repos, however we also
need to test the error path. We add a check for that as well by mocking
the fetch_cmd implementation of a GitRepo. We further need to remove the
commit, as otherwise the short-circuit implementation added in 021e961be
skips the fetching.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>