container-entrypoint: add custom git path to safedirs
The git clone path can also be controlled with the GIT_CLONE_PATH
environment variable. While this is ignored by kas (as the clone
is executed in a cleaned environment), it can happen when cloning the
main repository (the one kas then is run from). As the GitLab CI passes
the environment variables directly to the container invocation, the
GitLab CI cloning of the main repository already uses them, which fails
if the path is not added to the safedirs.

We now account for this special case.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>