container: Install binfmts for qemu manually
In trixie, qemu-user decided to exclusively rely on systemd for
installing its binfmt configs. First, we do not have and otherwise do
not need systemd in our container. Second, we cannot use systemd-binfmt
because it will always unregister pre-existing entries before
(re-)adding them. That could break parallel kas-isar runs because we can
still not benefit from binfmt namespaces with current isar.

Switch to a manual registration procedure instead which will not
overwrite existing entries, only add missing ones that are found in the
container's /usr/lib/binfmt.d folder and that are related to qemu-user.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>