Dockerfile: restore debian sources list after installation
We currently have the snapshot mirrors in the debian sources list. While
this in general is good to have a stable baseline for installing
additional packages, it is problematic in case the container gets older
and components become outdated (e.g. the ca-certificates). By restoring
the original debian mirror entries, we make it easier for our users to
keep things up to date by just running apt-update.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
[Jan: pushed source switching to the end of kas image build]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>